ECS: Change HashType from uint64_t to uint32_t and refactor hash functions - Updated HashType to uint32_t in ForwardDeclarations and Hash headers for consistency. - Refactored hash functions in Archetype to support both uint64_t and uint32_t, improving flexibility. - Removed unused hash specialization in Index header.

+++ src/libs/ecs/ForwardDeclarations.hpp
@@ -10,7 +10,7 @@ namespace RR::Ecs
 {
     template<int FixedSize>
     struct HashString;
-    using HashType = uint64_t;
+    using HashType = uint32_t;
     struct EntityId;
     struct Entity;
     struct Event;
@@ -26,7 +26,7 @@ namespace RR::Ecs
     struct ArchetypeEntityIndex;
     struct Archetype;
 
-    template <typename Tag, typename IndexType = size_t>
+    template <typename Tag, typename IndexType = uint32_t>
     struct Index;
 
     using EventId = Index<struct EventIdTag, HashType>;
